LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigate-event/event-constructor-expected.txt


Harness Error (TIMEOUT), message = null

PASS can't bypass required members by omitting the dictionary entirely
PASS destination is required
TIMEOUT signal is required Test timed out
TIMEOUT all properties are reflected back Test timed out
TIMEOUT defaults are as expected Test timed out
TIMEOUT hasUAVisualTransition is default false Test timed out
